LRQA-30742 Move to vm.provisioner package
